Remove code duplication via default implementation
This interface method is probably a case of YAGNI anyway. The only true
case where the behavior is different from clearing the reached set that
is passed as parameter is inside the ConditionAdjuster created via:
ConditionAdjusterFactories.STATIC.createConditionAdjuster

Here we do nothing instead, but the STATIC option is not used in a
configuration currently.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34786 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c