Allow distinction of blank edges from other CFA edges, even if edegs have the same endpoints.
This is necessary for the correct functioning of the combination of SlicingCPA and PredicateCPA.
The SlicingCPA introduces new blank edges for irrelevant CFA edges. If these edges later get relevant,
the original edges with meaning are used. Without a custom BlankEdges#equals method it is not possible to distinguish
the newly introduced blank edge from the original edge. Because of this, the cache for path formula creation fails.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/slicing-refinement@34879 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c