Fix concurrent modification in webclient script
The script used to occasionally throw the error:
"dictionary changed size during iteration"

If I am right this is because new file hashes are added to the cache
even after we initiated the shutdown where we want to write this cache
to disk. This commit essentially discards all new file hashes that
arrive after the shutdown was initiated, which is acceptable since
it is just a cache.

Fixes issue #466

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34994 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c