Fix handling of long double in predicate analysis
So far the long modifier was simply ignored,
now we apply the proper type size
(which is different for x86 and ARM).

Fixes #769

We might at some point move the information about exponent and mantissa
size to the machine model, but for this to make sense we probably also
need to restructure the machine model first
(ARM can also be a 32-bit Linux platform, so the name LINUX32 is
ambiguous).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@35076 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c