Add residual-program generation strategy 'REACHABILITY'
Strategy REACHABILITY writes to C the ARG created with an arbitrary analysis, as-is.
This allows to use `-setprop analysis.constructResidualProgram=true`
with any analysis to write the created ARG to C.

We also lessen the configuration check for analyses used with
residual program generation. This allows combination of condition-based
residual-program generation with arbitrary other analyses

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/byte_level_heap@35138 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c