Fix AttributeErrors when aborting webclient script
When the user requests a shutdown, the old behavior of the code was
to immediately stop the connection of the SSEClient object, but
this was actually still accessed concurrently to get the next
item. This commit causes the script to wait for the current item
to fully be processed and then terminates the connection and
makes sure we do not continue to iterate over the SSEClient.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/byte_level_heap@35167 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c