shortcut: avoid formula translation if we use identical solver instances.
This seems to fix a strange bug with MathSAT5
that appeared in the Buildbot [1] for commandlines like:

scripts/cpa.sh -predicateAnalysis -32 \
    -spec test/programs/benchmarks/properties/unreach-call.prp \
    test/programs/benchmarks/float-newlib/float_req_bl_0660a.c

I am not sure why it helps, because the translated formula can even be `true` and crashes MathSAT.
We might need further investigation here.

[1]
https://buildbot.sosy-lab.org/cpachecker/results/nightly-predicate/00174.nightly-predicate.diff.html

PS:
This commandline produces an UNKNOWN result.
Why do we export a witness for UNKNOWN anyway?

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@35353 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c