AutomatonCPA: fix matching for functioncalls.
Witness validation for concurrent programs seems to depend on our own renamed functions.
Thus, we need to also check for a match against the renmaed functions.

I tried to remove the code for creating thread-specific functionnames from the GraphML parser.
In theory, the GraphML parser does not need to know about our own renaming,
and we can simple match the original functionname during the analysis.
However, that results in an overhead of about factor three to five for the validation execution.
Thus, lets keep it that way.

This should also fix the failing WitnessExporterTest.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@35357 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c