Update ltl3ba binary
The binary requires the bdd-library 'BuDDy' to be executed sucessfully.
However, this led to errors in the previous version if the respective
file (libbdd.so) was not included in the variable 'LD_LIBRARY_PATH'.

I have now recreated the ltl3ba-binary in which the respective library
is now included and linked as a static library. It is therefore no longer
neccessary to have the library in the system of where cpachecker (and
hence the executable) will be deployed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/legion_dev@35411 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c