TypeHierarchy now saves JFieldDeclarations in an ImmutableSet
TypeHierarchy saved field declarations in a normal Set and thus unordered. Because of this, field declaration nodes were added to the cfa unsorted. This resulted in a bug while strengthening in test /sv-benchmarks/java/java-ranger-regression/apachecli_eqchk/
Because Options.java has many fields depending on each other, strengthen would try to strengthen on a field that would be earlier in the program, but later in our cfa.
CFAMethodBuilder#addNonStaticFieldMember sorts by starting line number of FileLocation before adding the fields now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@35466 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c