Webclient: fix script for Windows.
The fixed problem is the conversion of Windows/Linux path separators
that are transferred towards the VCloud webinterface.

Python itself can handle both separators ("/" and "\") on Windows to access files
(see docu for 'os.sep' and 'os.altsep').
Thus we can simply use the existing .xml files for benchmark definitions,
.set files for tasks collections, and .yml files for tasks from sv-benchmarks repository.

The VCloud web-service (and workers) can only handle the unix-style separator.
Thus we need to convert all filenames that are given towards the webinterface.

Note that this commit alone is not yet sufficient for running the script on Windows.
We still require the latest (not yet released) BenchExec version
that contains some fixes for Windows compatibility.

Current status: Using the script works on Windows:
- tasks are submitted via Webclient
- results are returned and stored
- the commandline for calling table-generator works fine

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@35613 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c