Windows script: improve parameter handling when using a equal sign.
We assume that the equal sign "=" only appears in the middle of the argument "-setprop X=Y"
The Windows script supports setting the heap and stack memory now.

With this change, all cmdlines that are executed on the Buildbot should be available on Windows,
by just replacing the executable 'cpa.sh' with 'cpa.bat',
as far as the native libraries are supported.

Thus, we can tell our students to just copy-and-paste cmdlines. :-)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36002 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c