Enable parsing of __attribute__(...)
A comment indicated that Eclipse CDT 8.1.1 had problems with that,
but we are using Eclipse CDT 9.4 by now and could upgrade to 10.0
if necessary.
integration-nightly-frontend shows no differences,
so it seems these old CDT problems have been fixed in the meantime.

Even older comments (from r654 (48f5077) and r698 (ac9844de) in 2009!)
indicated problems if parser.dialect is not set to GNUC,
but this is the default since r1180 (5070d20) from 2010.

So this should be safe, and some attributes change the code semantics,
so we actually need to parse them.
No handling for any attributes is added yet, but needs to come.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36019 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c