Handle __attribute__(mode(...)) for declarations
This is a GNU C extension that allows overriding the type of a
declaration, so it can change program semantics and we need to handle
it.
No changes in the analyses are needed, the frontend can take care of it.

Handles the most important part of #818
(those attributes that occur in common headers and change semantics).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36044 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c