Change callstack depth option for recursive termination analysis
The threshold of 10 was tested experimentally. Setting it to 10 reduces
the amount of wrong results by 7 using the SVCOMP-config, but has no further
effect on the outcome other than that.

Using a threshold lower than 8 also reduces the amount of correctly
solved tasks by 17-32, depending on the value of the config option
(here 2 and 4, respectively)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36059 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c