SMG: add support for extracting middle bits from a explicit value.
When reading from a region with an explicit value,
and accessing some bits in the middle of this region,
we can directly compute and return the correct value,
even without modifying the state.

Previously, in such cases a new symbolic value was introduced
and the explicit value of the region was deleted,
even if this is just a "read" operation.

Addin test task for this.

We should really think about rewriting some parts of SMG
to avoid repeated visits of the same expressions and
simplify debuging some of its confusing logic.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36082 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c