Temporarily restrict SMTHeapReadAndWriteTest to a single solver
On GitLab CI, this test nondeterministically fails with an abnormal JVM
exit. Typically, this is due to a OOM or a segfault.
However, sometimes the problematic test is listed as a Boolector test
although Boolector is always skipped for SMTHeapReadAndWriteTest.
Furthermore, I checked one instance of such a failure and there were
many GB of memory free. The tests are also always fast.
So, to investigate, let's just use one solver that is known to be stable
and see if the failures continue.

For this test, testing with several solvers is not important anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36095 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c