First part of supporting bit fields in SMTHeapWithByteArray
Part of #296

This now supports bit fields with up to 8 bits
if they happen to be aligned at a byte boundary.
Bit fields with more than 8 bits still cause an exception.
Bit fields that are not aligned at a byte boundary
are ignored currently anyway (independently of the heap model).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36140 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c