Do not filter states for further analysis in recursive BAM: leads to unsound analysis
If a target path that goes through a block A with a non-empty
waitlist in its reachedset,
and if there's some recursive function not exhaustivelly analyzed yet,
recursive BAM will do an additional analysis iteration.
In the new iteration, the now deactivated optimization
makes BAM only consider block-exits that were reachable from the states still in the
waitlist. If there are no new block exits (e.g., because of a non-terminating loop),
BAM will handle the block as if it had no exits at all,
and the previously found target location will not be found again.
This leads to unsound analyses.

Fixes #810.

In experiments on sv-benchmark's category ReachSafety-Recursive,
deactivating the optimization showed no negative effects.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36159 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c