Add bounded approximation for memcmp/strcmp/strncmp to PredicateCPA
Part of #664

Similarly to our implementation of strlen, we unroll these functions up
to a certain bound (by default set to 100).
Quantifiers could be used alternatively, but are not yet implemented.

This works with any kind of heap model as long as the pointers given
to these functions are always used as char pointers and were not cast
from other pointer types.
If the new byte-level heap from #296 is used, this works for all
pointer types.

Of course the performance of the deeply nested resulting formula
is not great and tasks that rely a lot on these functions will take
long, but at least we are sound unless the bound is reached.
For example, this helps to reduce several false alarms in the AWS
category of SV-COMP (cf. #656).

Although strcmp etc. are also used in other programs (e.g., LDV
drivers), it does not seem to cause a performance impact there.
On integration-predicate-bitprecise, no negative effect was noticeable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36162 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c