SMG: fix computation of memory cell boundaries.
That bug caused an invalid result in several uthash tasks, for example in
`uthash-2.0.2/uthash_JEN_test4-2.i`.

A minimal working (previously wrong) example is given with the following code.
The problem is/was an access to middle part of the second integer value in the struct.

```
typedef struct example_t {
    int id;
    int cookie;
} example_t;

void main() {
  example_t user = init();
  unsigned char *_hj_key = (unsigned char*)(&((user)->cookie));
  int x = ((unsigned)_hj_key[3] << 24);
}
```

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36166 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c