Remove resource limits and spec from partial configs
Touches only configs of which we are sure that removing
the specification is ok

Explanation mentioned in #838:
Config files in `config/components/` and `config/includes/`
are meant to be used or included by other configurations
instead of being used as stand-alone configs.
However, some of these partial config files define options
like specification or include resource-limits.properties,
which are default settings that should only be applied in
config files that are meant to be used stand-alone
(i.e., are in config/ itself).
Otherwise we get confusing and error-prone configs,
for exmaple if we want to include some config into an SV-COMP config,
which needs a different specification.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36206 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c