BMCAlgorithm: move the functionality of formula simplification into FormulaManager.
And adding a method to receive statistics about boolean operations.

Example task:

"""
scripts/cpa.sh -bmc-concurrency \
  -setprop cpa.loopbound.maxLoopIterations=10 \
  -setprop solver.solver=SMTInterpol \
  -setprop cpa.predicate.encodeBitvectorAs=INTEGER \
  -setprop cpa.predicate.encodeFloatAs=RATIONAL \
  -timelimit 300s -stats -32 \
  -spec test/programs/benchmarks/properties/unreach-call.prp \
  test/programs/benchmarks/pthread-wmm/safe033_pso.opt.i
"""

There is a speedup of factor 5 when using the option "-setprop bmc.simplifyBooleanFormula=true".
The optimization works for TRUE and FALSE tasks with SMTInterpol.
The overhead is minimal and the number of tasks with a slower runtime is small.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/ThreadingCPA-support-PredicateAnalysis@36400 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c