Add basic integration test setup for loopaccel
So far there is no way to quickly check wether the approach
still works. The idea is to have some small programs and
their expected verdict once the loops are abstracted.
Currently it seems we cannot run only the abstracted version
of the program, though this would be the analysis we would
like to run for the integration tests.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopaccel@36685 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c