Add two experimental options for imprecise counterexamples to predicate's path checker
alwaysUseImpreciseCounterexamples=true lets the path checker
always export an imprecise counterexample. If this option
is set, the path checker will never try to create a single
ARG path from the provided branching information, but
use the ARG path as-is (with merges).

This option must be used with allowImpreciseCounterexamples=true.

useSmtInfosForCounterexample=false tells the path checker
to NOT use the model of the SMT solver to make the counterexample
more precise.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariantWitness@36886 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c