Merge CI changes from branch report-js-build
For the developments in that branch we need some changes to the CI images
(newer node.js version) that bring other incompatible changes with them
(Chrome instead of Chromium due to new Ubuntu version).
This means that we need to update our Docker images
(which are used for all branches) together with the CI definitions.
So in order to be able to continue with the development in the branch,
we merge the minimal necessary set of changes for CI to trunk
and update our Docker images.
This change then also has to be merged to all other active branches,
otherwise the Javascript CI jobs would fail on them.
If everything goes as planned, the commands for doing this in a branch are

# start with clean SVN working copy on the respective branch
svn merge ^/trunk -c 36917
svn commit -m "Merge necessary CI adjustments for Javascript"

This should apply cleanly everywhere since the respective definitions
have not been changed for some time.

This merge also includes a single bug fix to the Javascript tests
and our Counterexample report regarding tooltips,
because with the new CI jobs it triggers a test failure
(probably because of a browser difference).


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@36917 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c