Make sure that predicateAnalysis-NoCEGAR writes the counterexample otherwise used for refinement
If no CEGAR is used at all, the predicate analysis will not write the counterexample
it computed with its ARG. Instead, the ARGStatistics will just take one path to the
found target state and create a counterexample from that.
For consistency, we want predicate analysis to output the (potentially infeasible)
counterexample that it would otherwise use for refinement, internally.

Do this with a new option that stops refinement early (and some hickhack)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariantWitness@36974 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c