Running -collect-undefined-functions should not output TRUE
UndefinedFunctionCollectorAlgorithm does not check anything actually,
so it should return NO_PROPERTY_CHECKED
(which was probably added in the meantime since the algorithm was written).

We could also make the analysis produce TRUE/FALSE
depending on whether there are undefined functions,
but that would be a bigger change that is not necessary right now.
So for now just avoid the misleading output.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37094 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c