Allow encoding ACSL annot. as unreach_call viol.
This assumes that reach_error() is already defined in the program,
so a proper mechanism to add this otherwise is still missing.
But for the ReachSafety category this should be fine.

This allows us to check the semantics of the ACSL annotations
with an arbitrary SV-COMP verifier instead. This is actually
now basically the same as MetaVal, only nicer because we don't
have to deal with the ugly generated residual programs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/acsl-bachelor@37113 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c