Fix bug in handling static variables
We convert these variables to global variables in the frontend,
but accidentally still used a name like for local variables
as the "qualified name".
So there were two different names for the same variable,
leading to confusion.
The end result was an assertion with
"Found imprecise counterexample in PredicateCPA."

This bug has likely existed since r7787 (5f9a8b84).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37195 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c