Fix bug in creating temporary variables for function pointers
Sometimes the type given by Eclipse CDT for an expression
is a function type (instead of a function-pointer type),
which would mean that we define temporary variables as functions
instead of function pointers.

This could lead to function pointers not followed properly,
or to the error message "Can't assign to functions" from
PredicateCPA's AssignmentHandler.
The latter occurred for example for
aws-c-common/aws_hash_table_{clean,clear,put,remove}_harness.yml.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37196 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c