Reduce log level of message about missing ARG
If we run an analysis without ARG, there should be no warning that the
ARG is not there, this would just be confusing.
In fact, we do not even print such a message for dumping the ARG itself,
so we also do not want it for dumping the witness view.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37198 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c