Fix crash in predicate abstraction with refinement selection
Fixes #884

During creation of the sliced infeasible prefixes,
the PredicateBasedPrefixProvider attempted to add SSA indices
to a block formula, which already has SSA indices.
This fails, but is an invalid operation anyways:
In a block formula, a range of indices can occur and be necessary
for the same variable (for example, for the statement "x = x+1;"),
but when we add indices to a formula, we add the same index to each
occurrence of a variable.

Before r25602 (a4746a6e), the code for adding SSA indices would have
attempted to replace all SSA indices in the block formula instead of
throwing an error and this would have resulted in a wrong formula
(with clauses such as x_2 = x_2 + 1),
but due to the PathFormula handling in PredicateBasedPrefixProvider,
only empty SSAMaps were passed, which made the operation a no-op
and returned the unchanged formula.
So this worked by chance.

That all SSAMaps were always empty in
PredicateBasedPrefixProvider.extractInfeasiblePrefixes can be seen from
the code that is removed by this commit:
It always started with an empty PathFormula and only conjuncted
BooleanFormulas to it, but this does not affect the SSAMap
(it cannot, because BooleanFormulas do not have this information).
In fact, the only ways to get a non-trivial SSAMap
are to get some passed in
(but this method only gets block formulas as BooleanFormulas, not PathFormulas)
or to create one iteratively to calling makeAnd(PathFormula, CFAEdge).

Code like "makeAnd(makeEmpty(pathFormula), booleanFormula)"
can be useful when wanting to add the SSA indices and pointer-target sets
from pathFormula to booleanFormula, but here it is equivalent to just
using booleanFormula.

So while I did not test it due to not having a suitable example right
now, I am reasonably sure that the code now does the same as originally.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37217 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c