Rewrite handling of solvers with partial theory support in PredicateCPA
For solvers that do not support certain theories,
we provide some unsound approximations using other theories.
But sometimes this is not enough, for example for solvers that support
only bitvectors there is no meaningful way to approximate floats.
Previously it was impossible to instantiate the class FormulaManagerView
(and thus all kinds of predicate analyses) for such solvers.

r32165 (a7bb1c9b) attempted to solve this by adding a configuration
option that would let the FormulaManagerView access the SMT theories of
the solvers lazily (only when actually requested by the analysis),
but this had several issues:
- The name createFormulaEncodingEagerly of the option did not describe
  what the effect of this option was in a way that was meaningful for
  users. In particular, it did not indicate that turning it off would
  put CPAchecker into a mode where crashes are expected depending on the
  input program.
- The option affected all theories at the same time,
  making it difficult to extend the same kind of handling to other
  theories.
- Due to related restructuring of the code, attempting an invalid
  approximation of theories led to a crash instead of a nice error
  message.
- Some other configuration options for theory approximations were now
  injected always instead of only when necessary, making our warning
  about unused configuration options less useful.

Thus this commit rewrites the code from r32165,
removes the above option, and provides a new way to achieve the same:
Each of the cpa.predicate.encode*As options for theory approximations
now as the additional possible value "UNSUPPORTED".
When setting this value, the FormulaManagerView will not attempt to use
this theory and instead will crash when the analysis tries to use it.
On startup a warning is printed that this may happen
(just like we do for theory approximations).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37245 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c