Conjunct ACSL loop invariant annotations per loop
If there are empty lines before the loop, we sometimes output two
loop invariant annotations. Frama-C doesn't like this and calls it a
"syntax error", though their ACSL spec doesn't explicitly forbid this.
Anyhow, we now just collect all invariants at a loop for as many empty
lines as there are and then generate one loop invariant annotation
where we build the conjunction of everything we collected.
We might want to do the same for the assertions later, but this is
not really a concern right now since Frama-C doesn't seem to bother.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/acsl-bachelor@37247 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c