Add annotations inside blocks, not between braces.
Annotations after a loop- or if-condition should not
be placed before the opening brace of the following block,
but afterwards.

Annotations are now potentially moved to a location where they do not hold
if the original program is formatted like so:
while (cond)
{ statement1;
statement2; }
Previously the annotation would have been placed after (cond), now it
will be placed after statement1.
This behavior is acceptable though, because generated programs are now syntactically
correct at least and we already require that there may not be multiple statements
on one line, we therefore simply extend this requirement, so that opening braces may
not be followed by another statement in the same line.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/acsl-bachelor@37251 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c