Remove d3 and dagre-d3 npm versions
Before, there were two versions of these libs that were used: the 
checked-in versions and the versions from npm. This removes all 
occurences of the npm version and only uses the checked in versions. To 
make this possible, the dagre-d3 file had to be updated to properly work 
without using npm. The new version is:

dagre-d3 v.0.6.4
https://dagrejs.github.io/project/dagre-d3/v0.6.4/dagre-d3.min.js

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/report-js-build@37289 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c