Remove PathFormulaManager.makeNewPathFormula(PF, SSAMap)
Reasons for this are the same as for the other overload of
makeNewPathFormula in the previous commit,
and in addition this method should not be used and was deprecated.
Most code that manipulates the SSAMap of a PathFormula will also
need to change the PointerTargetSet.
So far, we indicated this by having a deprecated overload of
makeNewPathFormula without the PointerTargetSet.
But this creates unnecessary duplication, especially given that now we
would need this in at least two places, as this is true for both
PathFormula.withContext() and
PathFormulaManager.makeEmptyPathFormulaWithContext().
So lets remove the deprecated overload without SSAMap.
This will also force callers to think about this question because now
they are forced to pass a PointerTargetSet and they need to get it from
somewhere.
The existing code is changed to have the same semantics as before
(i.e., no attempts at fixing the code that misses PointerTargetSet
handling has been made).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37379 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c