Make constructor of PathFormula package-private
There is at least one outside caller (CtoFormulaConverter)
that is supposed to create PathFormulas directly
(semantically is is part of the same package),
so we introduce a public factory method and mark it as deprecated
to discourage other callers.
(We do not mark the constructor as deprecated instead
to avoid having lots of deprecation warnings inside the same package.)

There are also three other remaining callers of the constructor
for which there is no easy other solution,
so these get to call the factory method for now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37381 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c