Make some classes in pathformula.pretty_print non-public.
These classes are part of an internal formula representation
that should not be used outside of this package
(because it is incomplete and does not cover all possible operators
and it is a mutable representation, so using it is error-prone).
In order to avoid people stumbling over it accidentally
(i.e., when using automcompletion) we make as much as possible
package-private.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37406 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c