Make mutating methods in pathformula.pretty_print non-public
Mutable data structures are error-prone and in general discouraged.
For example, the classes in this package implement equals() and
hashCode(), which indicates that they are intended to be put into Sets,
but this breaks if an object is changed after being added to the set.

As far as I see, no mutating operations occur on these instances
currently after they are returned from BooleanFormulaParser.parse(),
so there does not seem to exist an actual bug right now.
But to make sure this does not change we can and should make the
mutators package-private.

To make ExpressionNode safe, we also need to prevent outside access to
its internal list.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37407 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c