Added rudimentary witness2invariantWitness algorithm.
The algorithm converts from traditional graphml witnesses to the new
invariant witness format. The algorithm in the version of this commit is
rudimentary, because it does not extract any metadata from the witness.

The commit contains the algorithm, a run config and some model classes
that represent the new witnesses and their format. The witnesses are
written to yaml, thus jackson is added as a dependency.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariant-store@37414 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c