Let CFAToCTranslator keep skipped nodes
During CFA->C translation, join nodes are skipped to keep the hierarchy
of the generated C program as flat as possible: If all predecessors
of a join node are handled, we can avoid putting the code after the join
in one branch and using goto from the other.
But the implementation did not consider loop heads, which are a special
kind of join nodes because they have to be handled before all predecessors
can be handled - you can't (easily) handle a loop body before
handling the loop condition.
This missed characteristic lead to incomplete translations
if one branching (loop or if-else) and another loop followed each other.

We now respect loop heads.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37462 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c