Split java-runtime-licenses.txt into two files.
This file exists as documentation,
and as a handy way to let the reuse tool pick up all licenses of
dependencies. However, the reuse tool only scans the first 4kB
for license declarations, so we need to split the file into two files,
otherwise not all licenses are found.
This was not noticed before because there was a bug in the reuse tool
that prevented reporting of seemingly unused licenses.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37494 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c