Restructure handling of command-line arguments for Java
There is the old "CPAchecker" way of providing the main class as
command-line argument, and the new "SV-COMP" way of providing the paths
with source files as command-line arguments.
We support both and need to properly detect which one the user wanted.
This commit improves the respective heuristic.

There are two more changes:
- ".java" files can no longer be passed on the command line.
  This was yet another (only temporary) way and did not work for more
  than one file or files in packages.
- When the SV-COMP way is used, the config option for source path
  is not silently ignored but added to the paths from the command line.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@37523 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c