Attempt to fix problems with javascript-e2e-tests in CI
So far, we installed protractor globally and called webdriver-manager
from the global installation.
But when running the tests, npm would use the protractor installation
from the local node_modules, so the webdriver that was downloaded before
was not found.
Now we avoid the global installation and make sure to call the local
webdriver-manager.

Let's hope this does not create the same problem
that was attempted to be fixed with the global installation
in the first place (cf. branch fix-ci-js-2e2-tests, r36966 / 47fea77f).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37528 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c