bugfix to avoid modification of immutable set.
The bug only appears when verifying interpolants,
and that must be manually enabled via an option `cpa.predicate.refinement.verifyInterpolants`.

The bug appeared when debugging an issue with the newest version of Princess in JavaSMT `3.8.0-53-g502d1f0c`.
We get an invalid interpolant in some cases, causing repeated counterexamples. The issue is still unsolved.
Example:
  scripts/cpa.sh -predicateAnalysis -setprop solver.solver=PRINCESS \
    -setprop cpa.predicate.encodeFloatAs=INTEGER -setprop cpa.predicate.encodeBitvectorAs=BITVECTOR \
    -spec test/programs/benchmarks/properties/unreach-call.prp -32 \
    test/programs/benchmarks/loop-lit/gsv2008.c.i.v+lhb-reducer.c \
    -setprop cpa.predicate.refinement.verifyInterpolants=true

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37547 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c