Replace "catch Exception" with more specific catch in several cases
Cases where unchecked exceptions are immediately rethrown are ok,
and there are also a few cases where this happens inside the run() method
of a thread, were propagating unchecked exceptions would be better
but would require the creator of the thread to make sure they are
propagated.
But in other cases we do not want to catch unchecked exceptions because
they indicate bugs. In case of a bug it is also better to not attempt to
do some cleanup, as this might just cause further problems and
complicate debugging.
So we change the catch clause to only match checked exceptions.
In some cases, all checked exceptions were already handled by other
clauses, then the clause is removed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37566 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c