Move InvariantWitness related code to the resp package.
Creating invariantWitnesses from an ExpressionTree is a common task.
Thus, this algorithm is moved to the new class InvariantWitnessFactory.
Some visibilities were also fixed: The classes in InvariantWitnessWriter
are only dummy classes to define the yaml model. But linters dont like
them because the classes and their fields are never used in our code.
Adding getters silences the false warnings.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariant-store@37626 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c