Refactor InvariantGenerator interface result methods.
Before, there was a get method that returned AggregatedReachedSets.
This return value was always immediately wrapped by a FormulaSupplier or
ExpressionTreeSupplier in client code. Therefore, the get method is
replaced to directly return the supplier, like KIndGenerator already did.
This also causes the abstraction to not leak KIndGenerator anymore
(removed instanceof and casts).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariant-store@37628 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c