Rudimentary version of InvariantExportAlgorithm.
The algorithm basically mimics BMC with auxiliary invariants,
but exports them instead of using them for BMC. The goal is to
use the invariants for coop. verification (through the invariant store).

The algorithm is rudimentary because it is only a proof of concept impl.
Most missing parts are marked with todo comments in the code.
There is an example configuration (invariantExport.prop) for trying out
the new  algorithm.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invariant-store@37629 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c