Correct handling of if/else for block structure.
The post-state of an if-statement in the presence of an else
is not directly after the if, but only after the else.
It therefore makes sense to include both if and else in the
same block (as was already done before), but it has to be ensured
that the entering edges for both are considered part of the block.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/acsl-bachelor@37641 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c