Remove package cpa.ifcsecurity
The code was never developed further after being added in 2016.
In 2019 a branch for ifc has been created,
so it seems that if any code is still relevant,
it would be in that branch.
There are also no config files that use it,
so it was probably never used.

Keeping the package would mean additional effort
when we update static-analysis tools like Error Prone,
because these tend to find problems in the package.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37801 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c