Reformat all code related to handling Java programs
Typically, we do not want to reformat all code,
but for these packages, there are more advantages and less disadvantages
than for the rest of the code.
The reason is that this code is old and mostly untouched
in the last few years.
So the formatting of this code is often worse than in other parts of
CPAchecker, for example it even has wrong indentation sometimes,
and this gets more inconsistent every time new code is added
or existing code is changed.
Running "svn blame" is also less useful than otherwise due to the age of
the code.
And there is no ongoing work on this in any branches,
so there won't be any merge conflicts.
So now is a good time to reformat it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@37813 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c